sources/k8s: Fix nil/interface confusion with resourceVersion
If 'rawObj' was nil and it was type casted to 'obj', then
'obj' will not equal to nil as it's an non-nil interface value
with a pointer to a type, but with a nil inner pointer. Fix
the crash resulting from this by not using the type casted 'obj'
with resourceVersion().

Signed-off-by: Jussi Maki <joamaki@gmail.com>